Moved get_json to json module
